# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: REST API หลังบ้านเว็บไซต์โรงพยาบาลสะเมิง (ข่าว กิจกรรม บุคลากร ITA/MOIT
    ผู้ใช้งาน สิทธิ์เมนู)
  title: Hospital Samoeng API
  version: "1.0"
paths:
...
  /ita/years/{id}/moit:
    get:
      parameters:
      - description: year id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: list MOIT ของปีที่เลือก
      tags:
      - moit
    post:
      consumes:
      - application/json
      parameters:
      - description: year id
        in: path
        name: id
        required: true
        type: integer
      - description: ชื่อ MOIT
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: สร้าง MOIT ในปีที่เลือก (เฉพาะ admin)
      tags:
      - moit
  /items/{id}:
    delete:
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบ item (เฉพาะ admin)
      tags:
      - moit
    put:
      consumes:
      - application/json
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: integer
      - description: ชื่อใหม่
        in: body
        name: body
        required: true
        schema:
          properties:
            label:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไข item (เฉพาะ admin)
      tags:
      - moit
...
  /moit/{id}:
    delete:
      parameters:
      - description: moit id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบ MOIT (เฉพาะ admin, ลูก topic/item/ไฟล์หายตาม CASCADE)
      tags:
      - moit
    put:
      consumes:
      - application/json
      parameters:
      - description: moit id
        in: path
        name: id
        required: true
        type: integer
      - description: ฟิลด์ที่จะแก้
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไข MOIT (เฉพาะ admin)
      tags:
      - moit
  /moit/{id}/topics:
    get:
      parameters:
      - description: moit id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: list topic (พร้อม items) ของ MOIT
      tags:
      - moit
    post:
      consumes:
      - application/json
      parameters:
      - description: moit id
        in: path
        name: id
        required: true
        type: integer
      - description: ชื่อ topic
        in: body
        name: body
        required: true
        schema:
          properties:
            label:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: สร้าง topic ใน MOIT (เฉพาะ admin)
      tags:
      - moit
...
  /topics/{id}:
    delete:
      parameters:
      - description: topic id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: ลบ topic (เฉพาะ admin)
      tags:
      - moit
    put:
      consumes:
      - application/json
      parameters:
      - description: topic id
        in: path
        name: id
        required: true
        type: integer
      - description: ชื่อใหม่
        in: body
        name: body
        required: true
        schema:
          properties:
            label:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: แก้ไข topic (เฉพาะ admin)
      tags:
      - moit
  /topics/{id}/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: topic id
        in: path
        name: id
        required: true
        type: integer
      - description: ชื่อ item
        in: body
        name: body
        required: true
        schema:
          properties:
            label:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: สร้าง item ใน topic (เฉพาะ admin)
      tags:
      - moit
...
securityDefinitions:
  BearerAuth:
    description: ใส่ค่าเป็น "Bearer <token>" (ได้ token จาก POST /login)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"